requirements.txt
torch-efficient-distloss==0.1.3
tqdm==4.62.3
Pillow==9.1.0
pandas==1.4.0
opencv-python==4.6.0.66
imageio==2.21.1
matplotlib==3.5.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.8.0
tabulate==0.8.10
trimesh==3.10.2
tensorboard==2.11.0
tensorboardX==2.6
nerfacc==0.5.2
pyntcloud==0.3.1
ninja==1.11.1
open3d==0.16.0
lpips==0.1.4
pytorch-msssim==1.0.0